requirements.txt
speechrecognition
pyaudio
django
django-cors-headers
django-webpack-loader
djangorestframework
pygments